Proyecto finalizado
Proyecto completado. Faltaría comentar el código y ordenarlo para que sea más entendible

+++ src/Service/BorrarEscuderiaService.java
@@ -6,14 +6,19 @@ import lib.Escuderia;
 
 public class BorrarEscuderiaService {
 	public boolean checkDelete(Escuderia escuderia) {
-		
-		String sql = ("DELETE FROM escuderia WHERE nombreEscuderia = '"+escuderia.getNombreEscuderia()+"'");
-		
-		if(Conexion.EjecutarDeleteEscuderia(sql)) {
-			JOptionPane.showMessageDialog(null, "Escuderia eliminada");
-			return true;
-		}else {
+
+		if (escuderia.getNombreEscuderia().equals("Sin escuderia")) {
+			JOptionPane.showMessageDialog(null, "Esta escuderia no puede editarse");
 			return false;
+		} else {
+			String sql = ("DELETE FROM escuderia WHERE nombreEscuderia = '" + escuderia.getNombreEscuderia() + "'");
+
+			if (Conexion.EjecutarDeleteEscuderia(sql)) {
+				JOptionPane.showMessageDialog(null, "Escuderia eliminada");
+				return true;
+			} else {
+				return false;
+			}
 		}
-}
+	}
 }
